test(zwave_js): extract event factories into helpers module
